+++ src/com/saintsrobotics/corebot/output/TestBoardMotors.java
@@ -2,8 +2,8 @@ package com.saintsrobotics.corebot.output;
 
 public class TestBoardMotors extends Motors {
     
-    @Override public int getGearDropOut() { return -1; }
-    @Override public int getGearDropIn() { return -1; }
+    @Override public double getGearDropOut() { return -1; }
+    @Override public double getGearDropIn() { return -1; }
     
     public TestBoardMotors() {
         super(0, 2, 1, 3, 7, 6, 5, 4,
